Update README.md
The previously chosen twitter handle (textfiles) makes reading the example a little bit confusing (especially for novices) as it can be mistaken for part of the program syntax.

Even though I'm not a novice, I had to re-read the sentence a couple of times because I thought that the word textfiles was part of the command.

I picked Elon Musk just because is well known and has a not ambiguous name. Hope it makes sense!